refactor: move home_page logic to water_tracking_screen
